+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Added `CONTRIBUTING.md` file.
 - Add NURBS surface to PPM conversion examples.
 - Utilized ForUnitTest for testing.
+- Added Support for `OpenMP` and `do concurrent`.
 
 ### Changed